requirements.txt
pandas
selenium
pillow
requests
streamlit